Reorganized my code so that the code for the Time component is separate from sign up container, and the Time component gets called in the sign up container
