+++ src/components/pages/SignUp/SignUpContainer.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useHistory, Link } from 'react-router-dom';
 import Container from '@material-ui/core/Container';
 import RenderSignUp from './RenderSignUp';
-//import Time from './timeInZone';
+import Time from './timeInZone';
 import schema from './formSchema';
 //import axiosWithAuth from '../../../state/utils/axiosWithAuth';
 import axios from 'axios';
@@ -26,7 +26,7 @@ const initialUsers = [];
 const initialDisabled = true;
 
 //Time
-const API_KEY = 'JZN77L5UEHBN';
+//const API_KEY = 'JZN77L5UEHBN';
 
 const SignUpContainer = () => {
   //states
@@ -36,10 +36,9 @@ const SignUpContainer = () => {
   );
   const [disabled, setDisabled] = useState(initialDisabled);
   // const [formErrors, setFormErrors] = useState(initialFormErrors)
-  //TIme states
-  //const [timeData, setTimeData] = useState([]);
-  const [unixTimeStamp, setUnixTimeStamp] = useState();
-  //const [timeCountry, setTimeCountry] = useState();
+  //Time states
+  //const [unixTimeStamp, setUnixTimeStamp] = useState();
+
 
   //Helpers
   let history = useHistory();
@@ -84,51 +83,41 @@ const SignUpContainer = () => {
   //console.log(formErrors.name);
 
   //Time
-  useEffect(() => {
-    axios
-        .get(`http://api.timezonedb.com/v2.1/list-time-zone?key=${API_KEY}&format=json`)
-        .then(res => {
-            //console.log(res);
-            //setTimeData(res.data); 
-            //console.log(timeData);
-            // console.log(res.data);
-            // console.log(res.data.zones);
-            // console.log(res.data.zones[203]);
-            // console.log(res.data.zones[203].timestamp);
-            setUnixTimeStamp(res.data.zones[203].timestamp);
-            // console.log(unixTimeStamp);
-        })
-        .catch(err => {
-            console.log("ERROR:", err);
-        });
-  }, [unixTimeStamp]);
-
-  // const selectCountryTime = timeData => {
-  //     let country = timeData.filter(obj => obj.countryCode !== "ZA");
-  //     console.log("Country:", country);
-  //     setUnixTimeStamp(country.timestamp);
-  // };
+  // useEffect(() => {
+  //   axios
+  //       .get(`http://api.timezonedb.com/v2.1/list-time-zone?key=${API_KEY}&format=json`)
+  //       .then(res => {
+  //           //console.log(res);
+  //           //setTimeData(res.data); 
+  //           //console.log(timeData);
+  //           // console.log(res.data);
+  //           // console.log(res.data.zones);
+  //           // console.log(res.data.zones[203]);
+  //           // console.log(res.data.zones[203].timestamp);
+  //           setUnixTimeStamp(res.data.zones[203].timestamp);
+  //           // console.log(unixTimeStamp);
+  //       })
+  //       .catch(err => {
+  //           console.log("ERROR:", err);
+  //       });
+  // }, [unixTimeStamp]);
 
 //SK 1600931035 [422]
 //ZA 1600905835
 //KE [203]
 
-  //unixTime = Math.floor(Date.now() / 1000);
-  const convertTime = unixTimeStamp => {
-      let milliseconds = (unixTimeStamp * 1000) + (3600000 * 5); //3600000
-      let dateObject = new Date(milliseconds);
-      let timeCountry = dateObject.toLocaleString();
-      return (timeCountry);
-  };
+  // //unixTime = Math.floor(Date.now() / 1000);
+  // const convertTime = unixTimeStamp => {
+  //     let milliseconds = (unixTimeStamp * 1000) + (3600000 * 5); //3600000
+  //     let dateObject = new Date(milliseconds);
+  //     let timeCountry = dateObject.toLocaleString();
+  //     return (timeCountry);
+  // };
 
-  // selectCountryTime(timeData);
-  let timeCountry = convertTime(unixTimeStamp);
+  // let timeCountry = convertTime(unixTimeStamp);
 
   return (
     <div>
-      {/* <Time 
-        timeCountry = {timeCountry}
-      /> */}
       <Container maxWidth="sm">
         <RenderSignUp
           values={formValues}
@@ -137,11 +126,12 @@ const SignUpContainer = () => {
           formErrors={formErrors}
           disabled={disabled}
         />
+        <Time />
       </Container>
-      <div>
+      {/* <div>
         <h4>Date and Time in Kenya</h4>
         <p>{timeCountry}</p>
-      </div>
+      </div> */}
     </div>
   );
 };
