+++ src/components/pages/SignUp/timeInZone.js
@@ -2,45 +2,51 @@
 //http://api.timezonedb.com/v2.1/list-time-zone?key=JZN77L5UEHBN&format=json
 
 import React, { useState, useEffect } from 'react';
-// import axios from 'axios';
-// const API_KEY = 'JZN77L5UEHBN';
+import { makeStyles } from '@material-ui/core/styles';
+import axios from 'axios';
+const API_KEY = 'JZN77L5UEHBN';
 
-export default function Time (props) {
-    const { timeCountry } = props;
-    // const [timeData, setTimeData] = useState([]);
-    // const [unixTimeStamp, setUnixTimeStamp] = useState();
-    // const [timeCountry, setTimeCountry] = useState();
+const useStyles = makeStyles({
+    timeBlock: {
+        textAlign: 'center',
+    },
+});
 
-    // useEffect(() => {
-    //     axios.get(`http://api.timezonedb.com/v2.1/list-time-zone?key=${API_KEY}&format=json`)
-    //         .then(res => {
-    //             console.log(res);
-    //             setTimeData(res.data);
-    //         })
-    //         .catch(err => {
-    //             console.log("ERROR:", err);
-    //         });
-    // }, []);
+export default function Time () {
+    //Style
+    const classes = useStyles();
+    //Time states
+    const [unixTimeStamp, setUnixTimeStamp] = useState();
 
-    // const selectCountryTime = timeData => {
-    //     let country = timeData.filter(obj => obj.countryCode === "ZA");
-    //     console.log("Country:", country);
-    //     setUnixTimeStamp(country.timestamp);
-    // };
+    //useEffect and Axios
+    useEffect(() => {
+        axios
+            .get(`http://api.timezonedb.com/v2.1/list-time-zone?key=${API_KEY}&format=json`)
+            .then(res => {
+                setUnixTimeStamp(res.data.zones[203].timestamp);
 
-    // //unixTime = Math.floor(Date.now() / 1000);
-    // const convertTime = unixTimeStamp => {
-    //     let milliseconds = unixTimeStamp * 1000;
-    //     let dateObject = new Date(milliseconds);
-    //     setTimeCountry(dateObject.toLocaleString());
-    // };
+            })
+            .catch(err => {
+                console.log("ERROR:", err);
+            });
+      }, [unixTimeStamp]);
 
-    // selectCountryTime(timeData);
-    // convertTime(unixTimeStamp);
+    //SK 1600931035 [422]
+    //ZA 1600905835
+    //KE [203]
+
+    const convertTime = unixTimeStamp => {
+        let milliseconds = (unixTimeStamp * 1000) + (3600000 * 5); //3600000
+        let dateObject = new Date(milliseconds);
+        let timeCountry = dateObject.toLocaleString();
+        return (timeCountry);
+    };
+
+    let timeCountry = convertTime(unixTimeStamp);
 
     return (
-        <div>
-            <h4>Date and Time in South Africa</h4>
+        <div className = {classes.timeBlock}>
+            <h4>Date and Time in Kenya</h4>
             <p>{timeCountry}</p>
         </div>
     );
